# controllers/SMS_db/generated/StudentsControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE StudentsControllerGenerated.js PLEASE EDIT ../StudentsController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import StudentsModel from "../../../models/SMS_db/StudentsModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import StudentsController from "../StudentsController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/students`;
    router.post(baseUrl + "", authorize([]), StudentsController.create);
    router.delete(baseUrl + "/:id", authorize([]), StudentsController.delete);
    router.get(baseUrl + "/findBy_Class/:key", authorize([]), StudentsController.findBy_Class);
    router.get(baseUrl + "/findByclass/:key", authorize([]), StudentsController.findByclass);
    router.get(baseUrl + "/findByfirstname/:key", authorize([]), StudentsController.findByfirstname);
    router.get(baseUrl + "/findBylastname/:key", authorize([]), StudentsController.findBylastname);
    router.get(baseUrl + "/findByrollno/:key", authorize([]), StudentsController.findByrollno);
    router.get(baseUrl + "/:id", authorize([]), StudentsController.get);
    router.get(baseUrl + "/:id/get_Class", authorize([]), StudentsController.get_Class);
    router.get(baseUrl + "/:id/get_Exams", authorize([]), StudentsController.get_Exams);
    router.get(baseUrl + "", authorize([]), StudentsController.list);
    router.post(baseUrl + "/:id", authorize([]), StudentsController.update);
  },


  // CRUD METHODS


  /**
  * StudentsModel.create
  *   @description CRUD ACTION create
  *
  */
  create: async (req, res) => {
    try {
      const result = await StudentsModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * StudentsModel.delete
  *   @description CRUD ACTION delete
  *   @param String firstname First Name
  *   @param ObjectId id Id
  *   @param String lastname Last Name
  *
  */
  delete: async (req, res) => {
    try {
      const result = await StudentsModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * StudentsModel.findBy_Class
  *   @description CRUD ACTION findBy_Class
  *   @param Number class Class
  *
  */
  findBy_Class: async (req, res) => {
    try {
      const result = await StudentsModel.findBy_Class(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * StudentsModel.findByclass
  *   @description CRUD ACTION findByclass
  *
  */
  findByclass: async (req, res) => {
    try {
      const result = await StudentsModel.findByclass(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * StudentsModel.findByfirstname
  *   @description CRUD ACTION findByfirstname
  *
  */
  findByfirstname: async (req, res) => {
    try {
      const result = await StudentsModel.findByfirstname(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * StudentsModel.findBylastname
  *   @description CRUD ACTION findBylastname
  *
  */
  findBylastname: async (req, res) => {
    try {
      const result = await StudentsModel.findBylastname(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * StudentsModel.findByrollno
  *   @description CRUD ACTION findByrollno
  *
  */
  findByrollno: async (req, res) => {
    try {
      const result = await StudentsModel.findByrollno(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * StudentsModel.get
  *   @description CRUD ACTION get
  *   @param String firstname First Name
  *   @param ObjectId id Id resource
  *   @param String lastname Last Name
  *
  */
  get: async (req, res) => {
    try {
      const result = await StudentsModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * StudentsModel.get_Class
  *   @description CRUD ACTION get_Class
  *   @param Number class Class
  *
  */
  get_Class: async (req, res) => {
    try {
      const result = await StudentsModel.get_Class(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * StudentsModel.get_Exams
  *   @description CRUD ACTION get_Exams
  *   @param String examname Exam Name
  *
  */
  get_Exams: async (req, res) => {
    try {
      const result = await StudentsModel.get_Exams(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * StudentsModel.list
  *   @description CRUD ACTION list
  *
  */
  list: async (req, res) => {
    try {
      const result = await StudentsModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * StudentsModel.update
  *   @description CRUD ACTION update
  *   @param String firstname First Name
  *   @param ObjectId id Id
  *   @param String lastname Last Name
  *
  */
  update: async (req, res) => {
    try {
      const result = await StudentsModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;