# swagger.yaml
openapi: 3.0.0
info:
  title: SMS
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /lessons/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Lessons
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Lessons
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lessons"
    post:
      summary: CRUD ACTION update
      tags:
        - Lessons
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lessons"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Lessons"
  /lessons/findBy_Subjects/{key}:
    get:
      summary: CRUD ACTION findBy_Subjects
      tags:
        - Lessons
      parameters:
        - name: name
          in: path
          description: Name of Subject
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lessons"
  /lessons/findBy_Teachers/{key}:
    get:
      summary: CRUD ACTION findBy_Teachers
      tags:
        - Lessons
      parameters:
        - name: firstname
          in: path
          description: First Name
          required: true
          schema:
            type: string
        - name: lastname
          in: path
          description: Last Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lessons"
...
  /lessons/{id}/get_Subjects:
    get:
      summary: CRUD ACTION get_Subjects
      tags:
        - Lessons
      parameters:
        - name: name
          in: path
          description: Subject Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
  /students/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Students
      parameters:
        - name: firstname
          in: path
          description: First Name
          required: true
          schema:
            type: string
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
        - name: lastname
          in: path
          description: Last Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Students
      parameters:
        - name: firstname
          in: path
          description: First Name
          required: true
          schema:
            type: string
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
        - name: lastname
          in: path
          description: Last Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Students"
    post:
      summary: CRUD ACTION update
      tags:
        - Students
      parameters:
        - name: firstname
          in: path
          description: First Name
          required: true
          schema:
            type: string
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
        - name: lastname
          in: path
          description: Last Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Students"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Students"
  /students/findBy_Class/{key}:
    get:
      summary: CRUD ACTION findBy_Class
      tags:
        - Students
      parameters:
        - name: class
          in: path
          description: Class
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Students"
...
  /students/{id}/get_Class:
    get:
      summary: CRUD ACTION get_Class
      tags:
        - Students
      parameters:
        - name: class
          in: path
          description: Class
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
  /students/{id}/get_Exams:
    get:
      summary: CRUD ACTION get_Exams
      tags:
        - Students
      parameters:
        - name: examname
          in: path
          description: Exam Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    
  #-----  Subjects  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Lessons:
      properties:
        _id:
          type: string
        assignment:
          type: boolean
        class:
          type: number
        done:
          type: boolean
        lessonname:
          type: string
        lessonno:
          type: number
        subject:
          type: string
        teacher:
          type: string
        test:
          type: boolean
      required:
        -  _id
        - class
        - lessonno
        - subject
        - teacher
    Students:
      properties:
        _id:
          type: string
        class:
          type: number
        dateofbirth:
          type: integer
        fathersname:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        middlename:
          type: string
        mothersname:
          type: string
        percentageinlastexam:
          type: number
        rollno:
          type: object
      required:
        -  _id
        - class
        - firstname
        - lastname
        - rollno
...